requirements.txt
tk
matplotlib
pandas
numpy
seaborn
pillow
scikit-learn